// PIPREVENDA/frontend/src/types/api.generated.ts
export interface components {
    schemas: {
// ...
        ListarContratosResponse: {
            /** Format: int32 */
            id: number;
            contratoId: string;
            pcr: string;
            clienteNome: string;
            clienteCnpj: string;
            clienteNrSap: string | null;
            clienteUf: string | null;
            grupoEconomicoNome: string | null;
            segmento: components["schemas"]["Segmento"];
            tipo: components["schemas"]["TipoContrato"];
            /** Format: double */
            galonagemContratada: number;
            /** Format: double */
            galonagemFaturada: number;
            /** Format: double */
            galonagemPercentual: number;
            situacaoPcr: components["schemas"]["SituacaoPcr"];
            /** Format: date */
            fimVigencia: string;
            denuncia: boolean;
            bandeira: string;
            diretoria: string;
            regionalVendas: string;
            pontoVenda: string;
            consultor: string;
            papelGuardaChuva: components["schemas"]["NullableOfPapelGuardaChuva"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}